fix bugs, clean up all files, complete app opening structure
