+++ src/components/Categories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ScrollView, Text, StyleSheet, View, ImageBackground } from 'react-native';
+import { ScrollView, Text, StyleSheet, View, ImageBackground, TouchableOpacity } from 'react-native';
 
 const categories = [
   { id: '1', title: 'Community Life', description: 'Culture, Support, Assistance', background: require('../../assets/media1.jpg') },
@@ -8,28 +8,32 @@ const categories = [
   { id: '4', title: 'Health, Wellness', description: 'Physical, Mental Well-being', background: require('../../assets/media1.jpg') },
 ];
 
-const Categories = () => {
+const Categories = ({ onCategorySelect }) => {
   return (
     <ScrollView horizontal showsHorizontalScrollIndicator={false} style={styles.container}>
       {categories.map((category) => (
-        <ImageBackground
+        <TouchableOpacity
           key={category.id}
-          source={category.background}
-          style={styles.categoryItem}
-          imageStyle={styles.backgroundImage}
+          onPress={() => onCategorySelect(category.title)} // Pass selected category title
         >
-          {/* Overlay for the cover image */}
           <ImageBackground
-            source={require('../../assets/NewVersion/Cover.png')}
-            style={styles.coverImage}
-            imageStyle={styles.coverImageStyle}
+            source={category.background}
+            style={styles.categoryItem}
+            imageStyle={styles.backgroundImage}
           >
-            <View style={styles.textContainer}>
-              <Text style={styles.title}>{category.title}</Text>
-              <Text style={styles.description}>{category.description}</Text>
-            </View>
+            {/* Overlay for the cover image */}
+            <ImageBackground
+              source={require('../../assets/NewVersion/Cover.png')}
+              style={styles.coverImage}
+              imageStyle={styles.coverImageStyle}
+            >
+              <View style={styles.textContainer}>
+                <Text style={styles.title}>{category.title}</Text>
+                <Text style={styles.description}>{category.description}</Text>
+              </View>
+            </ImageBackground>
           </ImageBackground>
-        </ImageBackground>
+        </TouchableOpacity>
       ))}
     </ScrollView>
   );
