--- src/components/CreateButton.js
@@ -1,35 +1,0 @@
-import React from 'react';
-import { View, Button, StyleSheet, TouchableOpacity, Text } from 'react-native';
-import { withNavigation} from 'react-navigation';
-
-const CreateButton = ({ title, navigation }) => (
-    <TouchableOpacity 
-        onPress = { () => navigation.navigate('Todo')} 
-        style = {styles.buttonContainer}
-      >
-      <Text style={styles.buttonText}> {title} </Text>
-    </TouchableOpacity>
-  );
-  
-  const styles = StyleSheet.create({
-    buttonContainer: {
-      elevation: 8,
-      backgroundColor: "#C9D6DF",
-      borderRadius: 30,
-      paddingVertical: 10,
-      paddingHorizontal: 12,
-      alignSelf: 'center',
-      marginVertical: 40
-    },
-
-    buttonText: {
-      fontSize: 18,
-      color: "#000000",
-      marginVertical: 10,
-      marginHorizontal: 30,
-      alignSelf: "center",
-      fontFamily: "HelveticaNeue"
-    }
-  });
-
-  export default withNavigation(CreateButton);
