+++ src/components/ResultsList.js
@@ -1,120 +1,115 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, StyleSheet, FlatList, ImageBackground, TouchableOpacity, Image } from 'react-native';
+import React, { useState, useEffect, useCallback, useMemo } from 'react';
+import { View, Text, StyleSheet, ScrollView, ImageBackground, TouchableOpacity, Image } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
 const ResultsList = ({ title, results, navigation }) => {
-  if (!results.length) {
-    return null;
-  }
+  const [bookmarks, setBookmarks] = useState(new Set());
 
-  const getImageSource = (imageUrl) => {
-    if (imageUrl && (imageUrl.endsWith('.png') || imageUrl.endsWith('.jpg'))) {
-      return { uri: imageUrl };
-    } else {
-      return require('../../assets/icons/volunteer.png');
-    }
-  };
-
-  const BookmarkButton = ({ item }) => {
-    const [isBookmarked, setIsBookmarked] = useState(false);
-
-    const bookmarkKey = `@bookmark_${item.id}`;
-
-    const handleBookmark = async () => {
-      const newBookmarkStatus = !isBookmarked;
-      setIsBookmarked(newBookmarkStatus);
+  // 初始化书签数据
+  useEffect(() => {
+    const initializeBookmarks = async () => {
       try {
-        const bookmarkArray = JSON.parse(await AsyncStorage.getItem('@bookmarks')) || [];
-        if (newBookmarkStatus) {
-          const updatedBookmarkArray = [...bookmarkArray, item];
-          await AsyncStorage.setItem('@bookmarks', JSON.stringify(updatedBookmarkArray));
-        } else {
-          const updatedBookmarkArray = bookmarkArray.filter(bookmark => bookmark.id !== item.id);
-          await AsyncStorage.setItem('@bookmarks', JSON.stringify(updatedBookmarkArray));
-        }
+        const stored = await AsyncStorage.getItem('@bookmarks');
+        const parsed = stored ? JSON.parse(stored) : [];
+        setBookmarks(new Set(parsed.map(b => b.id)));
       } catch (e) {
-        console.error('Failed to update bookmark', e);
+        console.error('Bookmark initialization failed:', e);
       }
     };
+    initializeBookmarks();
+  }, []);
 
-    useEffect(() => {
-      const initializeBookmarkStatus = async () => {
-        try {
-          const bookmarks = JSON.parse(await AsyncStorage.getItem('@bookmarks')) || [];
-          setIsBookmarked(bookmarks.some(bookmark => bookmark.id === item.id));
-        } catch (e) {
-          console.error('Failed to initialize bookmark status', e);
-        }
-      };
+  // 书签切换逻辑
+  const handleBookmarkToggle = useCallback(async (item) => {
+    const newBookmarks = new Set(bookmarks);
+    const isExisting = newBookmarks.has(item.id);
 
-      initializeBookmarkStatus();
-    }, [item.id]);
+    isExisting ? newBookmarks.delete(item.id) : newBookmarks.add(item.id);
+
+    try {
+      await AsyncStorage.setItem('@bookmarks',
+        JSON.stringify([...newBookmarks].map(id => ({ id })))
+      );
+      setBookmarks(new Set(newBookmarks));
+    } catch (e) {
+      console.error('Bookmark update failed:', e);
+    }
+  }, [bookmarks]);
+
+  // 图片加载优化
+  const resolveImageSource = useCallback((uri) => {
+    const isValid = uri && /\.(jpe?g|png)$/i.test(uri);
+    return isValid ? { uri } : require('../../assets/icons/volunteer.png');
+  }, []);
+
+  // 渲染单个项目
+  const renderResultItem = useCallback((item) => {
+    const imageSource = resolveImageSource(item.image_url);
+    const isBookmarked = bookmarks.has(item.id);
+    const city = item.city || item.location?.city || 'Unknown City';
 
     return (
-      <TouchableOpacity onPress={handleBookmark} style={styles.bookmarkButton}>
-        <View style={styles.bookmarkIcon}>
-          <Ionicons
-            name={isBookmarked ? 'bookmark' : 'bookmark-outline'}
-            size={18}
-            color={'#884EFE'}
-          />
+      <TouchableOpacity
+        key={item.id}
+        style={styles.resultItem}
+        onPress={() => navigation.navigate('VolunteeringScreen', { itemData: item })}
+      >
+        {/* Top Part: Image with hours and bookmark */}
+        <ImageBackground
+          source={imageSource}
+          style={styles.backgroundImage}
+        >
+          <ImageBackground
+            source={require('../../assets/NewVersion/listCover.png')}
+            style={styles.coverImage}
+            imageStyle={styles.coverImageStyle}
+          >
+            <View style={styles.topContainer}>
+              <View style={styles.hoursContainer}>
+                <Text style={styles.hours}>{item.hours || 'Unlimited'} hours</Text>
+              </View>
+              <TouchableOpacity
+                onPress={() => handleBookmarkToggle(item)}
+                style={styles.bookmarkButton}
+              >
+                <Ionicons
+                  name={isBookmarked ? 'bookmark' : 'bookmark-outline'}
+                  size={18}
+                  color={'#884EFE'}
+                />
+              </TouchableOpacity>
+            </View>
+          </ImageBackground>
+        </ImageBackground>
+
+        {/* Bottom Part: Details */}
+        <View style={styles.detailsContainer}>
+          <View style={styles.cityContainer}>
+            <Image
+              source={require('../../assets/NewVersion/Location.png')}
+              style={styles.locationIcon}
+            />
+            <Text style={styles.city}>{city}</Text>
+          </View>
+          <Text style={styles.name}>{item.name}</Text>
+          <Text style={styles.organization}>{item.organization}</Text>
         </View>
       </TouchableOpacity>
     );
-  };
+  }, [bookmarks, handleBookmarkToggle]);
 
   return (
-    <View style={styles.container}>
-      <FlatList
-        showsVerticalScrollIndicator={false}
-        data={results.slice(0, 10)}
-        keyExtractor={(result) => result.id.toString()}
-        renderItem={({ item }) => {
-          const city = item.city || (item.location && item.location.city) || 'Unknown City';
-          const organization = item.organization ? item.organization : item.name;
-          const imageSrc = getImageSource(item.image_url);
-
-          return (
-            <TouchableOpacity
-              style={styles.resultItem}
-              onPress={() => navigation.navigate('VolunteeringScreen', { itemData: item })}
-            >
-              {/* Top Part: Image with hours and bookmark */}
-              <ImageBackground
-                source={imageSrc}
-                style={styles.backgroundImage}
-              >
-                <ImageBackground
-                  source={require('../../assets/NewVersion/listCover.png')}
-                  style={styles.coverImage}
-                  imageStyle={styles.coverImageStyle}
-                >
-                  <View style={styles.topContainer}>
-                    <View style={styles.hoursContainer}>
-                      <Text style={styles.hours}>{item.hours || 'Unlimited'} hours</Text>
-                    </View>
-                    <BookmarkButton item={item} />
-                  </View>
-                </ImageBackground>
-              </ImageBackground>
-
-              {/* Bottom Part: Details */}
-              <View style={styles.detailsContainer}>
-                <View style={styles.cityContainer}>
-                  <Image source={require('../../assets/NewVersion/Location.png')} style={styles.locationIcon} />
-                  <Text style={styles.city}>{city}</Text>
-                </View>
-                <Text style={styles.name}>{item.name}</Text>
-                <Text style={styles.organization}>{organization}</Text>
-              </View>
-            </TouchableOpacity>
-          );
-        }}
-      />
-    </View>
+    <ScrollView
+      showsVerticalScrollIndicator={false}
+      contentContainerStyle={styles.container}
+    >
+      {results.slice(0, 10).map(renderResultItem)}
+    </ScrollView>
   );
 };
+
+// 样式表保持不变...
 
 const styles = StyleSheet.create({
   container: {
@@ -126,28 +121,19 @@ const styles = StyleSheet.create({
     marginBottom: 20,
     marginHorizontal: 15,
     shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
+    shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.2,
     shadowRadius: 5,
     elevation: 5,
   },
-  backgroundImage: {
-    height: 100,
-  },
-  coverImage: {
-    height: 100,
-  },
-  coverImageStyle: {
-    opacity: 0.6,// Match the borderRadius of the categoryItem
-  },
+  backgroundImage: { height: 100 },
+  coverImage: { height: 100 },
+  coverImageStyle: { opacity: 0.6 },
   topContainer: {
-    justifyContent: 'space-between', // Ensures space between hours and bookmark
+    justifyContent: 'space-between',
     flexDirection: 'row',
-    alignItems: 'flex-start', // Aligns both at the top
-    padding: 10, // Adds padding inside the background
+    alignItems: 'flex-start',
+    padding: 10,
   },
   hoursContainer: {
     borderRadius: 6,
@@ -155,7 +141,7 @@ const styles = StyleSheet.create({
     paddingHorizontal: 5,
     paddingVertical: 3,
     height: 30,
-    justifyContent: 'center', // Centers text vertically
+    justifyContent: 'center',
   },
   hours: {
     color: '#884efe',
@@ -167,12 +153,6 @@ const styles = StyleSheet.create({
     backgroundColor: "rgba(255, 255, 255, 0.7)",
     width: 30,
     height: 30,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  bookmarkIcon: {
-    width: "100%",
-    height: '100%',
     alignItems: 'center',
     justifyContent: 'center',
   },
@@ -206,6 +186,5 @@ const styles = StyleSheet.create({
     fontWeight: '500',
   },
 });
-
 
 export default ResultsList;
